commit before sync
