+++ src/config/constant.js
@@ -1,0 +1,1 @@
+exports.DOMAIN = "http://localhost:5173";
