--- src/routes/stripeRoute.js
@@ -1,9 +1,0 @@
-const express = require("express");
-
-const { payment } = require("../controller/stripeController");
-
-const router = express.Router();
-
-router.post("/", payment);
-
-module.exports = router;
